Sort Contact Activities tab by Start Date
Mailchimp Activities will come into Acumatica in batch and will not
be brought in real time. We want to make sure they are displayed in
sequence with existing Acumatica activities and according to the actual time
the activity occurred not the time they were brought in Acuamtica.

To accomplish this, write the Mailchimp timestamp to the StartDate field
and then sort the Acitivities view by StartDate.