Hide "Add Mailchimp" from Add Activites Menu
On the Activities grid (on the Contacts screen and other screen with an
Activites tab), there is a button for Add Activity which has a dropdown
of buttons. These buttons are populated by any activity types which
are active and which have "IsInternal" set to true.

We don't want people to be able to manually add activities with the type
Mailchimp. Those should only come from real Mailchimp activities. So we
set "IsInternal" to be true on the Mailchimp Activity Type.

The meant we needed to change the restrictors on the Type field in the
CRActivity DAC. We want the options in the Type dropdown on this screen
to only include types which are active, not from the portal and not
mailchimp. But we also need to make sure we can still create activities
of the Mailchimp type with our Sync Mailchimp Activity action on the
Contact screen.

We wanted to handle all this with DAC attributes rather than some code
in the graph extension since the Activities grid appears on quite a few
screens and we don't want to have to repeat the code.