Enabling multiple actions on Mailchimp process screen
Jeremy had code that allowed you to select an action, but in this commit
we change the process delegate depending on the action selected

Also, on the Contacts screen, change the action that pulls in MailChimp
activities to be available to the process screen

Finally, I added a Delete MailChimp Activities button to make testing
easier. When the code is put live, that code should be commented out.