Upgrade to 2020 R2 - Restrictions on Activities changed
In the upgrade from 2019 R1 to 2020 R2, the PXRestriction in the base
code on CRActivities.Type changed. We are overriding the restrictions so we need to update our override

Objectives: Have the Mailchimp Activity Type be available for the code to create activities with but NOT be listed on the Add Activity dropdown on the Activities tabs of Business Accounts, Cases, etc. and also NOT be available to be chosen as a type from the Activity screen itself. 

Activity dropdown: It appears that if isSystem = false (isSystem is a calculated field and is true if the class is Event, Email or Task) and the application is "backend", then the activity type will show up in the Activity dropdown. Because we don't want to change is system, and because we can't set the application to "portal" (it won't let you create portal activities through the UI), we created a new application called "integration" (MailchimpAttribute.cs and EPActivityTypeExtensions.cs). If we use the integration application then the Mailchimp type is not listed on the Add Activity dropdown.

Activity Entry Screen type options: To get the Mailchimp type to not be an option when on the screen to add a new activity, wed added a new PXRestrictor to CRActivity.Type.  (CRActivityExtensions.cs)

If we only needed to add this restrictor, we wouldn't have to remove the base restrictors. But we also have to modify an existing restrictor so that there's not an error raised for types with the application "integration". Because we need to modify that, we have to remove base restrictors and re-add them. In some ways, this architecture is not ideal because we are overriding code rather than just adding to it. But for now, it seems like the best option. In the upgrade from 2019 R1 to 2020 R2, the PXRestriction base code on CRActivities changed, so we have to update our PXRestriction to match -- hence this commit.